Refactoring Skills
Part one is completed in introducing a new level of indirection. There
is still much to do to get things properly wired up but this is closer
to a fully testable model with the information most relevant to classes
being close to them instead of the behemoth that the charactersheet was
turning into.